feat: implement parameter column visibility toggle in PacketDataTable
