+++ app/src/app/components/contents/PacketGroup/utils/initializeColumnVisibility.ts
@@ -1,0 +1,7 @@
+export const initializeColumnVisibility = (parameterKeys: Set<string>) => {
+  const visibility: Record<string, boolean> = {};
+  parameterKeys.forEach((key) => {
+    visibility[`parameters_${key}`] = false; // hide all parameters by default
+  });
+  return visibility;
+};
